separate api in different controllers. One ApiController with 700 lines code does not provide clarity
